+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added: `git branchless init` now sets `advice.detachedHead false`, to reduce the incidence of scary messages.
 - Added: aliasing `git` to `git-branchless wrap` improves which commands are grouped together for `git undo`, and possibly enables more features in the future.
 - Added: Created `git move` command, which rebases entire subtrees at once.
+- Added: The `git-branchless wrap` command can now take an explicit `--git-executable` parameter to indicate which program to run.
 - Fixed: Visible commits in the smartlog sometimes showed the reason that they were hidden, even though they were visible.
 - Fixed: The working copy was sometimes left dirty after a `git undo`, even if it was clean beforehand.
 - Fixed: `git-branchless` now supports Git v2.31.